+++ src/pages/MapViewPage.tsx
@@ -9,7 +9,7 @@ import {
 } from "@/features/mapView/ui";
 import BackButton from "@/features/mapView/ui/BackButton";
 import { useEventStore } from "@/shared/stores";
-import { MapHeader } from "@/widgets/Header";
+import { MapHeader } from "@/widgets/Headers";
 import { AxiosError } from "axios";
 import { useEffect } from "react";
 import { useSearchParams } from "react-router-dom";
