+++ src/pages/VisitedPage.tsx
@@ -2,7 +2,7 @@ import FirstStep from "@/features/visited/ui/FirstStep";
 import SecondStep from "@/features/visited/ui/SecondStep";
 import Button from "@/shared/ui/Button";
 import StepIndicator from "@/shared/ui/StepIndicator";
-import { BackHeader } from "@/widgets/header";
+import { BackHeader } from "@/widgets/headers";
 import { useState } from "react";
 import { useNavigate } from "react-router-dom";
