+++ src/widgets/headers/model/types.ts
