+++ src/widgets/headers/ui/CloseHeader.tsx
